chore: organize root files
