+++ README.md
@@ -62,6 +62,7 @@ Run the test suite after installing dependencies:
 ```sh
 npm test
 ```
+Tests live alongside modules in `src/` and under `tests/` for DOM and sprite checks. 
 The test suite uses a jsdom environment with a stubbed Canvas context from [`jest.setup.js`](./jest.setup.js) and currently reports **42 passed suites (232 tests)** with coverage **84 % statements, 76 % branches, 84 % functions, 88 % lines**. Continuous integration runs the same command on each push. Detailed reports and UAT results are documented in `docs/04-test.md`.
 
 ## Versioning
