+++ tests/officeman-idle-sprites.test.js
