+++ tests/officeman-walk-sprites.test.js
