+++ tests/ol-idle-sprites.test.js
