+++ tests/ol-walk-sprites.test.js
