+++ tests/splash-screen.test.js
