+++ tests/student-idle-sprites.test.js
