+++ tests/student-walk-sprites.test.js
