+++ tests/trunk-move-sprites.test.js
